Implement modal
- TODO: Need more refactor regarding conditional rendering
