+++ packages/client/src/views/Settings/ChangePhonePage.test.tsx
@@ -18,7 +18,7 @@ import {
 import { createStore } from '@client/store'
 import { getStorageUserDetailsSuccess } from '@opencrvs/client/src/profile/profileActions'
 import { ReactWrapper } from 'enzyme'
-import { ChangePhonePage } from './ChangePhonePage'
+import { ChangePhoneModal } from '@client/views/Settings/ChangePhoneModal'
 import { changePhoneMutation } from '@client/views/Settings/mutations'
 import { queriesForUser } from '@client/views/Settings/queries'
 import { waitForElement } from '@client/tests/wait-for-element'
